Introduced sparse matrix struct
